Optimize Chain external objects.
Add FlatChain::FromExternal(), analogous Chain::FromExternal().
Remove Chain::{Append,Prepend}External().
Add Chain::{Append,Prepend}(FlatChain) instead.

Add FromExternal() overloads taking a tuple of constructor arguments for
the external object instead of an already constructed external object to
be copied or moved. This avoids constructing a temporary object and
copying or moving from it.

In FromExternal(), do not create an internal block when the external
object would be short. This is not feasible when the object is not
constructed yet, and this is actually not always faster. This should not
be a problem: when multiple tiny external objects are appended, they are
merged together by existing Append() code, and also many callers avoid
creating an external node if the data are short enough.

Add missing extern template declarations and definitions for
Chain::{Append,Prepend}Block() instantiations which are used in chain.h
but defined in chain.cc.

In fd_reader.cc, change MMapRef to avoid storing the pointer and
size which are already stored in the Chain block. Call munmap() by
operator()(string_view) instead of the destructor to access the data.

PiperOrigin-RevId: 254052586