Remove {,Backward}Writer::WriteSlow(string&&). Write(string&&) delegates to Write(Chain&&).
This will make easier to make Riegeli support both choices of
_GLIBCXX_USE_CXX11_ABI, which is required for open source TensorFlow.

Optimize Chain::AppendTo(string&&) if the Chain can be moved from and
consists of a string and the destination is empty. This restores the
optimization removed from StringWriter::Write(string&&).

Optimize out a string move in Chain::StringRef constructor. It is always
applied to an rvalue.

PiperOrigin-RevId: 254945291