Avoid dereferencing a null pointer if worker_ is nullptr in RecordWriterBase::Done().
This can happen if RecordWriter's destination was already unhealthy
(e.g. closed) during RecordWriter initialization.

This caused a mysterious crash in https://github.com/google/riegeli/issues/5

After this fix the behavior is still not too helpful (RecordWriter is
unhealthy with FailedPreconditionError("Object closed")), but it is not
obvious how to improve that, and at least there is no crash.

PiperOrigin-RevId: 263125330