Decrease kMaxBytesToCopy from 511 to 255.
In some cases this is faster, in some cases this is slower, but in
overall looks more like a win for important cases (like reading a Chain
from a ChainReader, BufferedReader, or FdMMapReader).

This increases the range of lengths which trigger the slow path
(virtual, potentially sharing) of Reader::Read() and Writer::Write().
There is still a possibility to tweak some method overrides to employ
copying if sharing has a high overhead for a particular type.

Also, tweak comments at kMaxBytesToCopy and Chain::kAllocationCost, and
move the definition of Chain::kMaxShortDataSize to the pre-C++17 section.

PiperOrigin-RevId: 263352580