Inside Chain methods, lower the threshold for copying from the generic kMaxBytesToCopy (255) to a new kMaxBytesToCopyToChain (kMaxShortDataSize, usually 16).
In contrast to Reader and Writer methods which use kMaxBytesToCopy,
in Chain methods the fast path is not inlined and the slow path is not
virtual, so there are fewer reasons to prefer copying.

Chain handles external objects quite efficiently. And even if the
external block is tiny, it will get rewritten if adjacent tiny blocks
appear.

PiperOrigin-RevId: 268902216