Make zstd compressor recycling more efficient.
If the input is flat and its size is known (i.e. the first
ZSTD_compressStream2() call is with ZSTD_e_end), zstdlib tunes
compression parameters for that size. Reflect this in
ZSTD_CCtxKey::size_hint_class. This requires creating the zstd
compressor lazily.

Also, avoid passing size_hint to zstdlib if final_size is known. This is
redundant for zstdlib and more costly than necessary.

PiperOrigin-RevId: 269508559