Let reading varints not move the cursor on failure instead of pointing the cursor past the failure.
This is analogous to strong exception safety.

While this does not matter for existing use cases, this is more useful
for reading from a growing source: when reading a varint fails, it can
be retried when the source grows.

The old way was the only thing possible before `Reader::Pull()` could
request `min_length > 1`, but now it can be done better.

PiperOrigin-RevId: 277911840