`IstreamReader` must not leave `eofbit` set on the `istream`.
It turns out that while `seekg()` clears `eofbit` before doing anything
else, other operations like `peek()` or even `tellg()` turn on `failbit`
if `eofbit` was set on entry.

In other words `peek()` is not idempotent. Assuming that the stream
pointer was at the end of the stream but we did not attempt to read
further yet:
- The first `peek()` call turns on `eofbit`;
- The second `peek()` call turns on `failbit` too.

And while `seekg()` succeeds after reading past the end of the stream,
`tellg()` fails.

I consider this design broken. `IstreamReader` must work around the
brokenness.

Since `Reader` assumes that reading past the end is not a failure,
`IstreamReader` clears `eofbit` after operations which may set it.
Also initially.

Also, fix some "ostream" → "istream" in failure messages.

PiperOrigin-RevId: 278810971