Allow `Writer::Push()` with `min_length > 1` even if writing that much data would be disallowed by `Writer::Write()`. Delay detecting overflow errors to synchronizing the cursor during future operations.
This affects `LimitingWriter` (overflow is checked by `SyncBuffer()`
instead of `MakeBuffer()`) and `ArrayWriter` (it uses `PushableWriter`).

Same for `BackwardWriter`. Add `PushableBackwardWriter` for that.

`WriteVarint{32,64}()` can now ask for `kMaxLengthVarint{32,64}` of
space instead of being careful to ask for minimal length.

`SnappyWriter` (via `WriterSnappySink`) and `FramedSnappyWriter` were
already calling `Push()` and then writing less data than space requested
(although using a compressor with a `Writer` with a limited capacity
would be a dubious practice so this was harmless).

PiperOrigin-RevId: 280237118