Pay more attention to the `recommended_length` parameter in `Reader::Pull()` and `{,Backward}Writer::Push()`.
The strategy is to return early if at least `min_length` can be
available without doing expensive work, but to allocate
`max(min_length, recommended_length)` if a new buffer needs to be
allocated.

Add a `recommended_length` parameter to
`PullableReader::PullUsingScratch()` and
`Pushable{,Backward}Writer::PushUsingScratch()`.

This improves the amortized efficiency of `Pull()` or `Push()` in a loop
which scans data for its terminator, e.g. when `min_length` is the
current length plus one (anything more might be sufficient) and
`recommended_length` is twice the current length (prefer to grow
geometrically).

In `Pushable{,Backward}Writer::SyncScratch()`, avoid losing the buffer
allocation if data from the buffer will be copied because at most
`kMaxBytesToCopy` was written.

Simplify `BufferedReader` and `tensorflow::FileReader`. Do not try so
hard to fill the allocated capacity of the buffer before allocating a
new buffer: just ask for the configured buffer size, delegating
efficiency concerns (overallocation, moving existing data back) to
`ChainBlock::AppendFixedBuffer()`. Call `ChainBlock::AppendSubstrTo()`
without special casing appending the whole buffer. OTOH these classes
can now make to a new buffer if existing buffer is filled sufficiently
but not completely, because this is simple enough now.

PiperOrigin-RevId: 282522242