Fix buffer sizing in `BufferedReader` (it was correct but inefficient).
The strategy is not obvious if a previous `ReadInternal()` did not fill
the whole buffer. The old logic was:
* Compute the recommended buffer length, based on the configured buffer
  size and depending on whether the size hint is near.
* If the amount of unconsumed data in the buffer is at least half of the
  recommended buffer length, append that data to the destination and
  make a fresh buffer using the recommended buffer length.
* Otherwise shift unconsumed data and fill the rest, up to the
  recommended buffer length in total.

The problem was that if the size hint is near, and if it was decided
to fill the rest of the buffer, then the amount of unconsumed data is
subtracted from the recommended buffer length twice. The other three
cases do not have this problem, so it cannot be fixed easily e.g. by
replacing `limit_pos()` with `pos()` during calculation of the
recommended buffer length (that would fix this case but would break
the case of making a fresh buffer).

Here is the new logic:
* Try to fill the rest of the buffer if this can be done without
  reallocation, but only up to the configured buffer size. This triggers
  when `ReadInternal()` did not fill the whole buffer (as long as the
  buffer is not shared with other Chains).
* If this could not be done, append unconsumed data to the destination
  and make a fresh buffer, computing the recommended buffer length for
  that.

The problem is avoided because the recommended buffer length is used
only in one of the cases.

In both the old and the new logic `PullSlow()` is special because it
cannot append unconsumed data to the destination.

For consistency, do the same in `tensorflow::FileReader`, even though it
does not have a size hint so it did not have this problem.

PiperOrigin-RevId: 294169373