In `BufferedWriter` and `tensorflow::FileWriter`, do not use the effective buffer size from previous operations, just the configured buffer size.
This makes the performance more predictable: a single `Push()` with
a huge `min_length` does not affect future operations.

In these classes, as well as in `FramedSnappyWriter`, include
`recommended_length` in buffer length calculation. This is less
problematic than before because it does not affect future operations.

Cosmetics: in these classes, as well as in `FramedSnappyWriter` and
`NullWriter`, apply the `Position` limit before resizing the buffer
instead of to the final buffer length, because there is no longer an
unpredictable increase of the buffer length caused by the memory
allocator, which needed to be trumped by the `Position` limit.

Cosmetics: in `FramedSnappyWriter`, call `BufferLength()` with
`min_length` of 1, not 0. It always returned a non-zero value if
`max_length > 0`, but this makes the intent clearer.
PiperOrigin-RevId: 294169974