Add overloads of `SerializeToWriter()` and `ParseFromReader()` with a const reference argument.
This allows to specify the template argument explicitly as `Writer*` or
`Reader*`, instead of `Writer*&` or `Reader*&` which would be needed to
use the forwarding reference overloads.

To reduce code duplication, move the implementations of these functions
to helper templates.

Manually inline the final portion of `internal::ParseFromReaderImpl()`
in `ParseFromChain()` to remove a redundant check for a flat source.

PiperOrigin-RevId: 294632882