Improve `SnappyWriter` memory usage by not trying so hard to maintain block boundaries which will avoid a copy inside the Snappy Encoder.
When e.g. appending another `Chain` consisting of small blocks to the
internal uncompressed `Chain` buffer, it can be better to share the
existing `Chain` blocks, even if their layout does not match Snappy
block boundaries, because the memory will be sitting around until
compression begins, possibly coexisting with the source `Chain`.

During compression only one block at a time is copied, which is not as
bad as copying the whole uncompressed source.

PiperOrigin-RevId: 295537530