Change the default brotli compression level from 9 to 6, and the default zstd compression level from 9 to 3.
This makes compression with the default level about 2 to 3 times faster
than before, with files larger by about 1% to 3%. Decompression speed is
similar.

The default zstd compression level matches the official zstd default.
The official brotli default is 11, but this is 100 times slower and 10%
denser, which is a good choice if compression density is much more
important than compression speed, but likely not as the default.

PiperOrigin-RevId: 299067540