In `{Chain,Cord}{,Backward}Writer`, do not bother returning the short buffer if `size_hint` is larger.
For `Chain` the change is done in `Chain::{Append,Prepend}Buffer()`
which used to return remaining short data whenever `min_length == 0`
because it promises to return whatever is available without allocation.
But it is only important to avoid making an allocation (so that clients
can rely on existing pointers not being invalidated); it is not needed
to return a non-empty buffer if there are reasons to avoid exposing it.

PiperOrigin-RevId: 302183337