Replace `Chain::kMaxBytesToCopyToChain` (16) with the bigger `kMaxBytesToCopy` (255).
Sharing blocks which are considered tiny is a poor choice if more than
one such block becomes adjacent, because they get rewritten and data is
copied anyway, while undoing an allocation. While it was faster to share
such a block alone, if another tiny block already precedes it, then it
was a clearly suboptimal decision, and it is hard to predict whether
another tiny block will follow it. It is simpler to defer to copying
that amount of data.

Replace `Chain::kMaxBytesToCopyFromCordToChain` with `kMaxBytesToCopy`
which has the same value (255).

PiperOrigin-RevId: 305202644