Add `absl::Cord&&` overloads to functions taking a `const absl::Cord&` parameter:
* `Chain::{Chain,Reset,Append,Prepend}()`
* `{,Backward}Writer::Write{,Slow}()`.
* `ChunkEncoder::AddRecord()`
* `RecordWriter::WriteRecord()`

This avoids doing an unnecessary atomic reference count adjustment.
For appending to a `Chain`, the optimization is possible only when the
`Cord` consists of a single node.

Make use of the overloads when possible by applying `std::move()` to the
argument of these functions if this is the last use of an lvalue.

Cosmetics: do not get `src.size()` as an argument in the same expression
as `std::move(src)`. It was harmless when the parameter is an rvalue
reference and there is no implicit conversion, but it would have been
wrong if the parameter was taken by value.
PiperOrigin-RevId: 306614521