Avoid making a buffer after `Writer` operations which bypass the buffer, if that might cause an allocation. The allocation is unnecessary if all operations during the `Writer` lifetime turn out to bypass the buffer.
But ensure that the buffer is not bypassed if the length is short, by
calling `WriteSlow()` from the base class which is implemented in terms
of `Push()`.

The cost of avoiding making a buffer is an additional virtual function
call if an operation which bypasses the buffer is followed by an
operation which does not.

Also, reset buffer pointers in `SyncBuffer()` and `PushInternal()`
methods. Since a `MakeBuffer()` call no longer necessarily follows, this
ensures that the buffer is resized if needed instead of leaving a stale
buffer for reuse.

Also, move `MaxBytesToCopyToCord()` to base.h, and avoid some
`BufferToCord()` calls if appending an `absl::string_view` is preferable
anyway.

Fix `CordBackwardWriter` inefficiency: the condition for using a short
buffer in `PushSlow()` was wrong (`start() == short_buffer_` instead of
`limit() == short_buffer_`) which caused it `Push()` to flush the short
buffer to the destination instead of copying it to the upgraded buffer.

PiperOrigin-RevId: 306614708