Remove `Tee{Writer,Reader}`.
They turned out to be not as useful as originally thought.

They are not an efficient way of achieving their goal: medium sized
writes/reads result in appending consecutive fragments of the same
`Chain` or `Cord` to the same destination, which leads to unnecessary
fragmentation.

If it is feasible, it can be better to remember the position range,
seek back, and append the whole range to another destination. This
avoids fragmentation, even if it is less general (less streamable).

PiperOrigin-RevId: 307361588