Optimize `{Write,Read}Varint{32,64}()` by handling the first byte outside of the loop.
Optimize `ReadVarint{32,64}()` by avoiding masking the highest bit of a
byte if it has been determined that the bit is unset. Otherwise it is
masked by subtracting a suitable constant in the next iteration.

Split `ReadVarint64()` in half, inlining only the first half, to
convince the compiler that the inlinable part is small enough to be
actually worth inlining and unrolling the loop.

Cosmetics: check whether varint is finished by comparing with 0x80
instead of bit masking.
PiperOrigin-RevId: 309003723