Avoid calling `MessageLite::ByteSizeLong()` unnecessarily.
This call is not free even if the size is needed again later. While
every message contains a cache of the size, `ByteSizeLong()` only writes
the cache, does not read it (to avoid having to invalidate the cache by
mutating operations). Only a separate `GetCachedSize()` function reads
the cache, but in general it is impossible to know whether the cache is
up to date.

Hence `SerializeOptions` is used to signal whether `GetCachedSize()`
is valid, and the code is restructured to call `ByteSizeLong()`
at most once. `SerializeToWriterImpl()` uses
`MessageLite::SerializeWithCachedSizes()` instead of
`MessageLite::SerializePartialToCodedStream()`.

As a side effect of this change,
`RecordWriterBase::WriteRecord(MessageLite)` and
`ChunkEncoder::AddRecord(MessageLite)` get an optional
`SerializeOptions` parameter, and `RecordWriterBase::Options`:
`metadata` and `serialized_metadata` are both optional,
without treating empty metadata specially.

PiperOrigin-RevId: 310193480