When `RecyclingPool<T>::global()` is called multiple times with different `max_size` arguments, the last `max_size` is in effect, instead of the maximum.
The motivation for the original rule was avoiding setting the size back
and forth if two pieces of code call `global()` for the same type but
with a different `max_size`. But such consensus should not be needed in
practice even in an unlikely scenario that different pieces of code use
the same type, and the current rule is simpler to understand.

PiperOrigin-RevId: 310903426