Use references instead of pointers for required output and input/output parameters.
Reorder parameters to place output parameters at the end more
rigorously, since we are losing one signal to highlight them.
They can still be followed by optional parameters.

Use references instead of pointers also for many other variables and
function results (not just function parameters) if they are never
`nullptr`.

Pointers are still used for optional output parameters.

Pointers are still used if the function in general needs the pointer
to be valid after the function returns. (This distinction used to be
meaningful only to input parameters.)

Pointers are still used in `Dependency<Ptr, Manager>`, i.e. if `Ptr` is
`T*`, then one of valid types for `Manager` is `T*`, and references
never enter this picture. In most of such cases they should be pointers
anyway because they need to be valid after the constructor returns;
the `Dependency` is a class member. Sometimes `Dependency` is used
in a function rather than a class and the pointer does not to be
valid afterwards (e.g. in `SerializeToWriter()`, `ParseFromReader()`,
`SnappyCompress()`), but accepting references here would be problematic
(they could use a 4th overload, but this would lead to 16 overloads
instead of 9 if there are two such parameters), hence it was left out.

`Std{In,Out,Err}()` are still pointers, to continue supporting the
automatic flushing behavior of `StdErr()`.

PiperOrigin-RevId: 312141877