Avoid making `Chain` substring blocks (`BlockRef`) referring to wasteful blocks, and avoid making `Cord` blocks referring to wasteful `Chain` blocks. Copy data instead in such cases.
This reduces the amount of memory pinned by the result of `Read()` to a
`Chain` or `Cord` from a class derived from `BufferedReader` under the
following conditions:

* total amount of data in the `Reader` is much smaller than its buffer
  size,
* a `size_hint` is not known.
* the result of `Read()` outlives the `Reader`,

at the cost of using **more** memory if the first two conditions hold
but the last condition does not hold.

While `Chain` avoids wasteful blocks other than the first or the last
one, this does not apply if the waste is hidden inside a substring
block, or in a `Cord`.

Appending a whole `ChainBlock` to a `Chain` still preserves a wasteful
block. This becomes the last block of the `Chain`, so the waste will be
reduced if more blocks are appended. Reducing the waste of the first or
last block of a `Chain` for long-lived and already finalized `Chain`s
will need some different design.

PiperOrigin-RevId: 313775314