Improve performance of `SnappyWriter` for medium-sized `string`s, `Chain`s, and `Cord`s: be more willing to copy data instead of sharing if the Snappy engine would otherwise copy them.
The Snappy engine copies data corresponding to each 64kB block if it is
not already flat. Hence data is shared only if it covers at least a
whole block (or if `size_hint` will be reached in the next block).

This can use more memory while data are being accumulated in the
`SnappyWriter` (a copy in `SnappyWriter` happens earlier than a copy in
the Snappy engine), but experiments indicate that this version is indeed
faster.

Also, inlike the only usage of `MakeBuffer()`, and make that computation
slightly more readable, using `RoundUp()` instead of `%` operator on
negated unsigned numbers.

PiperOrigin-RevId: 314515895