Fix propagating stream status in `SerializeToWriter()` and `ParseFromReader()`, this time correctly.
The inner implementation function already propagated stream status if
proto serialization or parsing failed. But even if proto parsing claims
to have succeeded, the proto might have been only partially parsed if
the stream failed, because `ZeroCopyInputStream` does not distinguish
failures from the end of stream.

Hence the responsibility for propagating the stream status if
`!healthy()` is always in the implementation function. The wrapper
template only adds propagating the status if `!Close()` in case the
stream is owned. In either case a stream failure takes precedence over
a parsing failure.

For consistency the same is done for serialization.

The previous fix left an incorrect `is_owning()` conditions in the
wrapper functions. This is now moot.

PiperOrigin-RevId: 316437900