In `ZstdReader`, avoid a memory copy when decompressing everything to a flat buffer.
This can make decompression up to 4% faster and using less memory
(for compression level 1, decompressing 32KB).

This is possible with `ZSTD_d_stableOutBuffer` option introduced in
zstd-1.4.5.

Unfortunately this optimization makes impossible to decompress as much
as possible from a truncated source, then retry when the source has
grown. Hence `growing_source` option is added to continue supporting
this behavior when needed.

PiperOrigin-RevId: 317436117