Fix a bug in seeking to the end of file which coincides with a block boundary.
In `ChunkReader::SeekToChunk()`, if the file ends at a block boundary,
then a chunk can be assumed to end there too. If this position is the
seek target (rather than the nearest block boundary before the seek
target), then searching for the chunk is done.

Cosmetics: in other places in `SeekToChunk()` where a block header or
chunk header could not be read, return early if `!healthy()`. This is
cleaner, avoids relying on `Fail()` doing nothing if a failure is
already set.

Cosmetics: write `pos_ < new_pos` instead of `pos_ <= new_pos`
(where `pos_ == new_pos` was already handled before), for consistency
with other conditions which check whether `new_pos` was reached or not
reached yet.
PiperOrigin-RevId: 321082440