Rolling this out again after a compiler fix.
Support CTAD (class template argument deduction) since C++17, e.g.:
- `ChainWriter writer(&chain);`
- `RecordReader reader(std::move(file_reader), options);`
or even:
- `auto writer = ChainWriter(&chain);`
- `auto reader = RecordReader(std::move(file_reader), options);`

Explicit deduction guides make clear for which classes CTAD is
supported, avoiding triggering `-Wctad-maybe-unsupported` when CTAD is
used without an explicit deduction guide.

CTAD is not used in Riegeli itself yet, because it is not supported in
C++14.

CTAD is disabled for constructors from tuples of dependency arguments,
because in this case there is no clue what should be the right template
argument, and just using the default is often inappropriate (passing a
tuple of dependency arguments is most useful for dependencies held by
value which are expensive to move, but the defaults are usually
pointers; having a different default for CTAD would be confusing).

CTAD is not disabled however for constructors which open a file from a
filename, because in these cases the default is good.

Also, deduction is no longer disabled via `internal::type_identity_t<T>`
for constructors from already open files, because in this case the
default makes sense (and was already supported).

PiperOrigin-RevId: 326305619