Rename `LengthLimitSetter` to `LengthLimiter`.
This leads to a more natural name of a variable which holds it:
`limiter`.

Weaken an invariant in a comment to match the reality: after
`set_size_limit()` which enlarges the size limit, `limit()` and
`limit_pos()` might not be as large as possible. They are only
guaranteed to not exceed the source limit nor the size limit.

(Changing the code to match the previous invariant instead might not be
the best idea, because that would require calling `src_reader()`.)

Add a private `LimitingReaderBase::reset_size_limit()`, a cheaper
variant of `set_size_limit()` when the limit is guaranteed to not
shrink. It is used by `LengthLimiter`.

PiperOrigin-RevId: 329930503