Add `ZlibReaderBase::Options().set_concatenate()` to decode concatenated compressed streams to concatenateion of their decompressed contents.
This is the behavior of the `gunzip` command line utility, and this is
specified in https://tools.ietf.org/html/rfc1952:

```
2.2. File format

A gzip file consists of a series of "members" (compressed data
sets).  The format of each member is specified in the following
section.  The members simply appear one after another in the file,
with no additional information before, between, or after them.
```

`ZlibReaderBase::truncated()` is available only if `concatenate()` is
`false`.

Remove `FramedSnappyReaderBase::truncated()`: it was a mistake because
it was unreliable, it was `false` between chunks.

Fix reading from a growing suurce in `FramedSnappyReader` by moving the
cursor only after the complete chunk is available.

PiperOrigin-RevId: 333573849