Improve Zlib status messages.
Zlib sets `stream->msg` inconsistently: sometimes to a detailed message,
sometimes to a translation of the return code, and sometimes not at all.

If it is not set, ensure that at least a translation of the return code
is included in the status message.

Also, in `ZlibReader`, set the status ccode to `kInternal` instead of
`kDataLoss` for errors during decompression which should not happen.

PiperOrigin-RevId: 334234546