PREPARATION FOR AN INCOMPATIBLE CHANGE in `riegeli::{String,Chain,Cord}Writer` and `riegeli::{Chain,Cord}BackwardWriter`.
The existing behavior of their constructor and `Reset()` is to append to
the destination. The new behavior will be to replace the destination,
unless `Options::set_append(true)` is given.

For `BackwardWriter`s, substitute “prepend” for “append”.

Rationale. The replacing semantics is simpler to reason about than the
appending semantics, because code has fewer preconditions: existing
contents of the destination do not matter.

Call sites belong to three families:

1. Replace the destination (this is currently done with an explicit
   `Clear()` before using the `Writer`).
2. Append to the destination.
3. Either semantics works because the destination is initially empty.

This change will treat 3 as a special case of 1 instead of as a special
case of 2.

This is consistent with how `Reader::Read()` was changed to replace the
destination instead of appending.

It is tempting to base APIs on the appending semantics, because it can
express the replacing semantics with an explicit `Clear()` before. But
it makes the intent less clear.

The default semantics continues to be appending until all callers are
migrated by adding `set_append(true)` when needed.

PiperOrigin-RevId: 336030362