WARNING: AN INCOMPATIBLE CHANGE in `riegeli::{String,Chain,Cord}Writer` and `riegeli::{Chain,Cord}BackwardWriter`.
Change the default value of the `append` option from `true` to `false`.

The behavior of `Options::set_append(false)` is temporarily to `CHECK`
that the destination was empty. This helps to verify that no code will
suddenly change the meaning. In a few days the behavior will be changed
to replace the current contents of the destination.

All users which relied on the old appending behavior have been changed
to specify `Options::set_append(true)`.

PiperOrigin-RevId: 337262809