Add `Object`-like API to stream classes not derived from `Object`: `WriterOstream` and `ReaderIstream`.
To implement that, move the implementation of the basic `Object` state (the
closed bit, and the failed bit with associated `absl::Status`) to an
`ObjectState` class, so that it does not require inheritance.

This adds or reimplements:
* `{WriterOstream,ReaderIstream}Base::close()`
  (not `Close()` so that it follows `std::{i,o}fstream`)
* `{WriterOstream,ReaderIstream}Base::healthy()`
* `{WriterOstream,ReaderIstream}Base::is_open()`
  (not `closed()` so that it follows `std::{i,o}fstream`)
* `{WriterOstream,ReaderIstream}Base::status()`
* `{WriterOstream,ReaderIstream}::Reset()`

Previously `WriterOstream` and `ReaderIstream` maintained their own `closed()`
state, but `failed()` and `status()` were delegated from the underlying `Writer`
or `Reader`. After this change the whole state is stored locally, which means
that after `close()` the underlying `Writer` or `Reader` can be freely modified,
without disturbing the `status()` of the `WriterOstream` or `ReaderIstream`.

Other changes:
* Mention that `Close()` may also synchronize the current position to the
  source or destination.
* Fixed `{WriterOstream,ReaderIstream}Base::status()` to report the original
  status when the object is failed and closed.
* Move `status()` implementation to `object.cc`. It is normally only accessed on
  failure paths so inlining it makes machine code longer for no benefit.

PiperOrigin-RevId: 341377191