Simplify the `Buffer` contract: remove the aspect of remembering the size to be allocated later. Also, avoid pathologic performance for repeated allocations with a slightly growing capacity (by at least doubling the capacity).
Originally buffer sizes were mostly static, but since the introduction of the
`min_length` parameters of `Push()` and `Pull()`, the declared buffer size is
stored aside anyway, and the buffer is potentially reallocated before each use.
This way storing the size for lazy allocation lost its utility.

And originally the declared buffer size was returned by `size()`, but now it
includes allocation overhead and the contract got more complicated.

Cosmetics in `NullWriter`: use the full allocated capacity of the buffer.

PiperOrigin-RevId: 345230649