Promote private `FlatCord()` function in `chain.cc` to a public `MakeFlatCord()` function in `base.h`.
Include a check which makes a regular flat `Cord` node if it is short enough.
The check is redundant in the context of usages in `chain.cc`, but not
otherwise.

Use `MakeFlatCord()` in `Buffer::ToCord()` for slightly cheaper allocation (does
not need to store the pointer and length again).

There will be another usage soon.

PiperOrigin-RevId: 345572780