Replace `Csv{Writer,Reader}Base::Options::set_standalone_record()` with standalone functions `WriteCsvRecordToString()` and `ReadCsvRecordFromString()`.
They are simpler to use. While they are limited to writing/reading to/from a
`std::string` instead of an arbitrary `Writer`/`Reader`, this should be
sufficient in practice.

Also, in `{Record,Csv}Reader`, clarify that when the recovery function returns
`false`, the behavior is as if the end of source was encountered. This requires
a special implementation in `ReadCsvRecordFromString()` because even an empty
source yields a record (a recovery function there is probably not useful, but
let it behave consistently for completeness).

PiperOrigin-RevId: 350041979