When `ReadLine(absl::string_view&)` returns `false` at the end of the source, clear `dest` as the comment promises.
On failure before a line terminator, let `ReadLine()` return `false`, because
the line might be truncated. But set `dest` to the partial line, in case the
caller prefers to see all data from the source.

This lets `ReadLine()` and `CsvReader` use the same policy for picking lines
(except that `CsvReader` throws away a partial line).

PiperOrigin-RevId: 355598219