Add `DigestingWriter` and `DigestingReader`.
They resurrect former `TeeWriter` and `TeeReader`, while changing the interface
of the side output to be more suitable for a checksumming algorithm. Rationale:
checksumming in this context does not need its own buffer (because it is
performed on data already buffered by the main output or input, so fragments are
not expected to be small), checksumming does not need to take ownership of
`Chain`s or `Cord`s being written, and checksumming does not yield failures.
Hence a simpler interface than `riegeli::Writer` suffices.

PiperOrigin-RevId: 355599104