Remove the precondition of `CsvReader::last_line_number()`: if no record was attempted to be read, return 1.
The precondition was erroneously treated as being violated if `ReadRecord()` was
attempted but `CsvReader` started as unhealthy due to the underlying `Reader`
being unhealthy.

This allows treating `last_line_number()` as an informative “current position”
in a setting where the next record is being buffered (hence `last_line_number()`
and not `line_number()`), without having to treat “we returned early because the
file could not even be opened” as a special case.

Also, ensure that `line_number()` is 1 even for a `CsvReader` which started as
closed.

PiperOrigin-RevId: 355821621