If the last line is not terminated by a newline, set `line_number()` as if it was.
While a line with that number arguably does not exist, and the current position
is still in the previous line, the current position is now after the previous
line rather than before it; this is the only case between `ReadRecord()` calls
when the current position is not the beginning of a line. It is less surprising
to claim that the current position is in the next line, than to suggest that
reading the last record did not move the current position.

PiperOrigin-RevId: 355824728