Add `CsvHeader` and `CsvRecord` to support reading/writing records with fields accessed by name.
New options `CsvReaderBase::Options::set_read_header()` and
`CsvWriterBase::Options::set_header()` cause the header to be read/written
automatically.

If a header is set, records being read/written can be expressed as `CsvRecord`
as an alternative to a sequence of field values.

Cosmetics: when writing a record expressed as any iterable of
`absl::string_view`s, constrain its type with `std::enable_if`, so that the
overload does not just accept any type, with build failures in the function body
if the type does not have the expected property.
PiperOrigin-RevId: 357556974