Fix `ambiguous template specialization` errors from GCC in `CsvHeader` (should be: ambiguous explicit template instantiation), by constraining the overload meant for multiple parameters to never match a single parameter.
Do this also for other variadic templates which overlap with a non-variadic
overload, even when an explicit template instantiation is not used — for
consistency, and to avoid relying on ordering of overloads; only a single
overload matches every call.

Also, add `extern template` declarations in header files for templates defined
in implementation files (with a single instantiation).

PiperOrigin-RevId: 358857318