Simplify conversion from `absl::string_view` to `std::string`, resolving a TODO.
The workaround is still needed for assigning `absl::string_view` to
`std::string`. It was never needed for conversion, I remembered wrongly.

Explanation:

C++17 introduced `std::string_view`, added an explicit constructor of
`std::string` from `std::string_view`, and an assignment operator of
`std::string` from `absl::string_view` (where there is no concept of whether it
is explicit).

In C++14 mode, `absl::string_view` is a distinct type which emulates that
behavior as much as possible: it provides an explicit conversion operator of
`absl::string_view` to `std::string`. But it cannot add an assignment operator
of `std::string` (it could have been emulated with implicit conversion, except
that this would enable unintended conversions as well, and would be less
efficient because does not reuse the existing `std::string` storage).

Hence Riegeli avoids assigning `std::string` from `absl::string_view`, but
relying on explicit conversions is fine.

This simplifies the implementation and improves the performance of creating
`CsvHeader` from names which are not already `std::vector<std::string>`. By
utilizing the `std::vector` constructor from two iterators, the `std::vector`
can be preallocated if iterators support that. I thought that this is not
feasible (except by some convoluted code) because converting individual elements
requires more explicit code, but actually it does not.

PiperOrigin-RevId: 358860888