Change the type of `CsvRecord::fields()` from `const std::vector<std::string>&` to `absl::Span<const std::string>` (or `absl::Span<std::string>` if the `CsvRecord` is not const), to add the non-const overload without allowing to break the invariant that the number of fields matches the header.
Change the type of `CsvHeader::names()` from `const std::vector<std::string>&`
to `absl::Span<const std::string>`, for consistency with `CsvRecord::fields()`
and to avoid a static empty vector just to return a reference to.

Add `CsvHeader::IndexOf()`, and deprecate indexing `CvsRecord` by
`CsvHeader::iterator`. Indexing `CsvRecord` by `CsvHeader::iterator` (to avoid a
string lookup each time) can be replaced by indexing the result of
`CsvRecord::fields()` by the result of `CsvHeader::IndexOf()`. The problem with
indexing by `CsvHeader::iterator` was that it required the iterator to belong to
the same `CsvHeader` as the `header()` of the `CsvRecord` being indexed, which
included copies with shared payloads, but did not include independent copies
(or copies made by a failed `CsvHeader::TryAdd()` which also makes the payload
unique). This is too subtle for a precondition, especially when CSV headers can
be amended in a function, and given that `operator==` does not distinguish
whether payloads are shared. Indices also can go out of sync, but at least they
yield an easier mental model for realizing when this happens.

PiperOrigin-RevId: 359494749